tree_enhanced_dl/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Tree models
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging (optional)
tensorboard>=2.13.0
wandb>=0.15.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
